frontend/requirements.txt
gradio>=4.0
Pillow
numpy
requests
python-dotenv
